+++ src/components/Compose/ComposeSend.tsx
@@ -11,10 +11,8 @@ import { Tooltip } from '@/components/Tooltip.js';
 import { MAX_POST_SIZE_PER_THREAD } from '@/constants/index.js';
 import { measureChars } from '@/helpers/chars.js';
 import { classNames } from '@/helpers/classNames.js';
-import { createMockComment } from '@/helpers/createMockComment.js';
 import { getCurrentPostLimits } from '@/helpers/getCurrentPostLimits.js';
 import { isValidPost } from '@/helpers/isValidPost.js';
-import { refreshComments } from '@/helpers/refreshComments.js';
 import { useIsMedium } from '@/hooks/useMediaQuery.js';
 import { useSetEditorContent } from '@/hooks/useSetEditorContent.js';
 import { ComposeModalRef } from '@/modals/controls.js';
@@ -38,17 +36,7 @@ export function ComposeSend({ post }: ComposeSendProps) {
 
     const [{ loading }, handlePost] = useAsyncFn(async () => {
         if (posts.length > 1) await crossPostThread();
-        else {
-            const updatedPost = await crossPost(type, post);
-            if (type === 'reply') {
-                const source = post.availableSources[0];
-                const parentPostId = post.parentPost[source]?.postId;
-                if (parentPostId) {
-                    const mockPost = createMockComment(source, updatedPost);
-                    refreshComments(source, parentPostId, mockPost);
-                }
-            }
-        }
+        else await crossPost(type, post);
         ComposeModalRef.close();
     }, [type, posts.length > 1, post]);
